[REM] Portado a funcao get_ocurrences para dentro do modelo de holidays
